Update instructions for Tracon 2026
