+++ client/src/markdown/KonstiInstructions.mdx
@@ -10,6 +10,8 @@ import { LoginLocalKompassi } from "./components/LoginLocalKompassi";
 import { SignupLottery } from "./components/SignupLottery";
 import { Expand } from "./components/Expand";
 import { HelpDesks } from "./components/HelpDesks";
+import { HasHelpDesks } from "./components/HasHelpDesks";
+import { NoHelpDesks } from "./components/NoHelpDesks";
 import { AppRoute, ProgramTab, ProfileTab } from "client/app/routes";
 
 <Fi>
@@ -59,14 +61,15 @@ Konsti is used to sign up for tabletop role-playing games, larps and workshops.
 
 <Tracon>
 <Fi>
-Konstin kautta voi ilmoittautua pöytäroolipeleihin ja kirpputorille.
+Konstin kautta voi ilmoittautua pöytäroolipeleihin, työpajoihin ja kirpputorille.
 </Fi>
 
 <En>
-Konsti is used to sign up for tabletop role-playing games and the flea market.
+Konsti is used to sign up for tabletop role-playing games, workshops and the flea market.
 </En>
 </Tracon>
 
+<HasHelpDesks>
 <Fi>
 Konstin käytössä auttavat tapahtuman vapaaehtoiset seuraavilla tiskeillä:
 </Fi>
@@ -76,6 +79,17 @@ Help with using Konsti is available from convention volunteers at these desks:
 </En>
 
 <HelpDesks />
+</HasHelpDesks>
+
+<NoHelpDesks>
+<Fi>
+Konstin käytössä auttavat tapahtuman vapaaehtoiset.
+</Fi>
+
+<En>
+Help with using Konsti is available from convention volunteers.
+</En>
+</NoHelpDesks>
 
 <Fi>
 Voit selata <Link to={ProgramTab.PROGRAM_LIST}>ohjelmalistaa</Link> kirjautumatta sisään, mutta ohjelmaan ilmoittautumiseen vaaditaan käyttäjätili.
@@ -201,12 +215,17 @@ Konsti accounts from previous events have been removed, so you will need to crea
 
 <LoginKompassi>
 <Fi>
+Aikaisempien tapahtumien Konsti-tunnukset eivät ole enää voimassa.
+
 Kompassi-tunnusta ei poisteta ellet ole erikseen sitä pyytänyt, joten jos olet aiemmin kirjautunut Konstiin Kompassi-tunnuksilla, voit nyt käyttää niitä uudelleen.
 
 </Fi>
 
 <En>
+Konsti accounts from previous events are no longer active.
+
 A Kompassi account is not deleted unless you specifically request it, so if you have previously logged in with a Kompassi account, that same account still exists and you can use it again.
+
 </En>
 </LoginKompassi>
 
@@ -316,6 +335,7 @@ You can sign up to program items in a lottery. After the lottery, remaining or c
 Ohjelmanumeroihin ilmoittaudutaan joko arvonnalla tai suoraan ilmoittautumisjärjestyksessä. Arvonnan jälkeen jäljelle jääneet paikat ja peruutuspaikat täytetään ilmoittautumisjärjestyksessä.
 
 - **Roolipelit**: Suora ilmoittautuminen
+- **Työpajat**: Suora ilmoittautuminen
 - **Kirpputori**: Arvonta
 
 </Fi>
@@ -324,6 +344,7 @@ Ohjelmanumeroihin ilmoittaudutaan joko arvonnalla tai suoraan ilmoittautumisjär
 You can sign up for program items either through a lottery or directly in sign-up order. After lottery, remaining or cancelled seats are filled in sign-up order.
 
 - **Role-playing games**: Direct sign-up
+- **Workshops**: Direct sign-up
 - **Flea market**: Lottery
 
 </En>
@@ -560,19 +581,19 @@ On Saturday, games starting between 10:00 and 12:00 use direct sign-up, as there
 <Expand>
 
 <Fi>
-### Roolipeleihin ilmoittautuminen
+### Roolipeleihin ja työpajoihin ilmoittautuminen
 </Fi>
 
 <En>
-### Role-playing game sign-ups
+### Role-playing game and workshop sign-ups
 </En>
 
 <Fi>
-Roolipeleihin ilmoittaudutaan suoraan eli paikat täytetään ilmoittautumisjärjestyksessä. Roolipeleihin ilmoittautuminen aukeaa 4 tuntia ennen pelin alkua.
+Roolipeleihin ja työpajoihin ilmoittaudutaan suoraan eli paikat täytetään ilmoittautumisjärjestyksessä. Ilmoittautuminen aukeaa 4 tuntia ennen ohjelmanumeron alkua, poislukien ohjelmanumerot, jotka alkavat ennen klo 12. Niihin ilmoittautuminen aukeaa edellisenä päivänä klo 18.
 </Fi>
 
 <En>
-Role-playing games use direct sign-up meaning that seats are reserved in sign-up order. Sign-up for role-playing games opens 4 hours before the game begins.
+Role-playing games and workshops use direct sign-up meaning that seats are reserved in sign-up order. Sign-up opens 4 hours before the program item begins, except for program items beginning before 12:00. For those, sign-up opens the previous day at 18:00.
 </En>
 
 </Expand>
@@ -773,7 +794,7 @@ Sign-up for other program items opens 4 hours before the program item begins, ex
 </En>
 
 <Fi>
-Traconin kirpputorille pääsee ostoksille pe klo 16–22 ja la klo 9:30–14 ainoastaan ajanvarauksella. Ilmoittautuminen tapahtuu arvontaa käyttäen ja jäljelle jääneet paikat ja peruutuspaikat täytetään ilmoittautumisjärjestyksessä. Arvontaan ilmoittautuminen alkaa perjantaina klo 08:00. Arvonnat tapahtuvat perjantaina kello 14:00 ja 17:00 sekä lauantaina kello 7:30.
+Traconin kirpputorille pääsee ostoksille pe klo 16–21:30 ja la klo 9:30–14 ainoastaan ajanvarauksella. Ilmoittautuminen tapahtuu arvontaa käyttäen ja jäljelle jääneet paikat ja peruutuspaikat täytetään ilmoittautumisjärjestyksessä. Arvontaan ilmoittautuminen alkaa perjantaina klo 08:00. Arvonnat tapahtuvat perjantaina kello 14:00 ja 17:00 sekä lauantaina kello 7:30.
 
 Kirpputorille ilmoittaudutaan jonotusaikoihin eli mihin aikaan voit saapua jonottamaan kirpputorille pääsyä.
 
@@ -782,7 +803,7 @@ Esitä Konstista löytyvä pääsylippu, kun saavut jonottamaan.
 </Fi>
 
 <En>
-At the Tracon flea market, shopping is possible on Friday from 16:00–22:00 and on Saturday from 9:30–14:00 only with a reservation. Signing up is done by lottery, and the remaining and cancelled spots will be filled in order of sign-up. Sign-up for the lottery begins on Friday at 08:00. The lotteries will take place on Friday at 14:00 and 17:00, and on Saturday at 7:30.
+At the Tracon flea market, shopping is possible on Friday from 16:00–21:30 and on Saturday from 9:30–14:00 only with a reservation. Signing up is done by lottery, and the remaining and cancelled spots will be filled in order of sign-up. Sign-up for the lottery begins on Friday at 08:00. The lotteries will take place on Friday at 14:00 and 17:00, and on Saturday at 7:30.
 
 You sign up for the flea market by choosing a queuing time — that is, the time you can arrive to line up for entry.
