+++ client/src/markdown/components/HelpDesks.tsx
@@ -1,29 +1,10 @@
 import { ReactNode } from "react";
 import { useTranslation } from "react-i18next";
-import { config } from "shared/config";
-import { EventName } from "shared/config/eventConfigTypes";
-
-interface HelpDesk {
-  nameFi: string;
-  nameEn: string;
-}
-
-const helpDesksByEvent: Record<EventName, HelpDesk[]> = {
-  [EventName.ROPECON]: [
-    { nameFi: "Larp- ja Roolipelitiski", nameEn: "Larp & RPG Desk" },
-    { nameFi: "Pelitiski", nameEn: "Gaming Desk" },
-    { nameFi: "Info", nameEn: "Info Desk" },
-  ],
-  [EventName.HITPOINT]: [{ nameFi: "Roolipelitiski", nameEn: "RPG Desk" }],
-  [EventName.TRACON]: [{ nameFi: "Roolipelitiski", nameEn: "RPG Desk" }],
-  [EventName.SOLMUKOHTA]: [
-    { nameFi: "Ohjelmatiski", nameEn: "Programme Desk" },
-  ],
-};
+import { getHelpDesks } from "./helpDeskList";
 
 export const HelpDesks = (): ReactNode => {
   const { i18n } = useTranslation();
-  const helpDesks = helpDesksByEvent[config.event().eventName];
+  const helpDesks = getHelpDesks();
   const isFinnish = i18n.language === "fi";
 
   return (
