+++ playwright/tests/about.spec.ts
@@ -44,15 +44,12 @@ test("Switch language between English and Finnish", async ({
   await languageSelector.selectOption("fi");
   await expect(aboutPage.heading("Osallistujille")).toBeVisible();
 
-  // Expand the help question: the desk names come from the HelpDesks component
-  // and are per-event, so match the Finnish word for a desk rather than a name
+  // Expand the help question: its answer names the help desks only for events
+  // that have them, so assert the part that is there either way
   await page.getByRole("button", { name: /Mistä saan apua/ }).click();
-  await expect(
-    aboutPage.main
-      .getByRole("listitem")
-      .filter({ hasText: /tiski$/ })
-      .first(),
-  ).toBeVisible();
+  await expect(aboutPage.main).toContainText(
+    "Vapaaehtoisemme auttavat Konstin käytössä ja ohjelmanumeroihin ilmoittautumisessa",
+  );
 
   // And back to English
   await languageSelector.selectOption("en");
